ref #518 move context to compile
